requirements.txt
logging==0.5.1.2
numpy==1.17.2
pandas==1.1.2
scipy==1.3.1
paddle.paddle==2.4.0
